Experiment with grammar-based recognition
There's no documentation on this grammar param as far as I can tell, but
it appears to be a JSON-encoded array of sample sentences, from which a
sort of finite state machine is derived at runtime. When used it
provides much more precise results, but may be tricky to get right.

https://github.com/alphacep/vosk-api/issues/55 suggests the possibility
of building the grammar offline based on a JSGF input, which would be
*ideal*. We would probably have to include that in the model.tar.gz
rather than using this runtime grammar mechanism, which is fine but
harder to manage than doing it in code...