Don't necessarily pause mic processing right when requested
If we do this, we will starve the processor, and will never get the
result. So instead, if we have requested to pause while a result is
outstanding we will wait for that result to be available and pause the
mic at that point. While waiting, we *disable* microphone input by
marking :voice as :busy?. (NOTE: we don't reject any events from this
state... yet...)