Rewrite nasr text parsing *again*, by hand, using okio
If we wanted to just skip through to the right position by hand-scanning
for the icao, we can get to KJFK in *just ~200ms*. Even fully-parsing
all apt and rwy records in-between, we can parse through KJFK in just
~2s. Attempting to do the same in gloss *never finished*. Heck, even
parsing the *very first record in the file* took 2s with gloss, which
makes no sense to me.

We should pull this parsing stuff out into a library and publish it.