Make find-airport-data an order of magnitude faster with type hinting
It was ~2.5s to read KJFK, which was already blazingly fast compared to
gloss, but it's now ~750ms which is a *massive* improvement.

This commit also adds record length data, which lets us do cool things
like skip parsing fields that we don't care about, resulting in the
`search-for-fixed-record` function which, while complicated to use, is
*fast*.

Basically it will read frame-length chunks at a time and skip *directly*
to the position of a particular field in a record, parse that field, and
pass it to a predicate. When that predicate returns `true`, it stops and
returns a sequence of Frames (`okio.Buffer` instances) corresponding to
that record size. This sequence can then be transformed however you like
(IE: if you want to use a different record to actually transform things)
with the normal functions.

By using this function, we can actually cut down the processing time for
KJFK to ~300ms, which is barely slower than the hand-optimized scan
loop.

`search-for-fixed-record` originally brought down the time from ~2.5s to
~1.5s. That was pretty nice, but I didn't quite believe that the lazy
sequence was still adding a whole order of magnitude of latency, so I
did more experimentation and the real silver bullet was type-hinting.
Tossing a `^Buffer` type hint on the params that interact with one
brought that 2.5s down to the 750ms for the method without
`search-for-fixed-record`--pretty remarkable!