Add tool to generate airport data from NASR records (#8)
* Experiment with reading AIXM formatted data

Sadly things like the navaids in the area, sids/stars, etc. are not
included in AIXM format yet. But the XML was very slow to parse anyway
so maybe this is a blessing.

* Experiment with using gloss to parse the NASR text files

This is *painfully* slow for some reason. Just reading a ~1500 byte
buffer off of a file takes ~1s at the repl which seems insane. I don't
understand what gloss is doing under the hood to be so slow

* Rewrite nasr text parsing *again*, by hand, using okio

If we wanted to just skip through to the right position by hand-scanning
for the icao, we can get to KJFK in *just ~200ms*. Even fully-parsing
all apt and rwy records in-between, we can parse through KJFK in just
~2s. Attempting to do the same in gloss *never finished*. Heck, even
parsing the *very first record in the file* took 2s with gloss, which
makes no sense to me.

We should pull this parsing stuff out into a library and publish it.

* Rewrite main tool implementation to use the nasr data parsing

* Pull okio layer into its own ns

* Make find-airport-data an order of magnitude faster with type hinting

It was ~2.5s to read KJFK, which was already blazingly fast compared to
gloss, but it's now ~750ms which is a *massive* improvement.

This commit also adds record length data, which lets us do cool things
like skip parsing fields that we don't care about, resulting in the
`search-for-fixed-record` function which, while complicated to use, is
*fast*.

Basically it will read frame-length chunks at a time and skip *directly*
to the position of a particular field in a record, parse that field, and
pass it to a predicate. When that predicate returns `true`, it stops and
returns a sequence of Frames (`okio.Buffer` instances) corresponding to
that record size. This sequence can then be transformed however you like
(IE: if you want to use a different record to actually transform things)
with the normal functions.

By using this function, we can actually cut down the processing time for
KJFK to ~300ms, which is barely slower than the hand-optimized scan
loop.

`search-for-fixed-record` originally brought down the time from ~2.5s to
~1.5s. That was pretty nice, but I didn't quite believe that the lazy
sequence was still adding a whole order of magnitude of latency, so I
did more experimentation and the real silver bullet was type-hinting.
Tossing a `^Buffer` type hint on the params that interact with one
brought that 2.5s down to the 750ms for the method without
`search-for-fixed-record`--pretty remarkable!

* Switch find-airport-data to be the "faster" version

* Make clojure-lsp linter happy

* Read airport elevation, magnetic variation, and position

* Refactor to automatically download current nasr data, read from zip

Unzipping would be slightly faster, but keeping it in a zip is kinda
nice....

* Fix and expand apt properties extraction

* Prepare basic parsing of FIX.txt file

Sadly we can't just use artcc, because eg MERIT is under ZBW. We'll
probably need to do like a distance search based on fix coordinates...

* Remove (broken?) protocol type hints to use coord-distance in tool

Perhaps type hints only work for concrete Java interfaces/classes. When
left on, we cannot require this ns due to an "Unable to resolve
classname" error. We *may* be able to get the generated interface for
these protocols, but it may not be worth it....

* Support filtering fixes by chart types

Using SID and STAR charts seems to work pretty well, although we do get
some overlap, eg VADDR is like *right* on top of LGA, and openscope does
not include it for KJFK....

* Add fix selection by ID

* Prepare more of a proper CLI

* Add STAR/SID parsing

For other files we've mostly left the original data as-is, but this file
is a bit... unintuitive to understand what to do with the raw records,
so we clean it up into something more natural

* Experimentally, use STARs/SIDs to select fixes

* Include departure procedure exit fixes

Looking pretty good!

* Include navaids from procedures

* Add --write flag to write data to disk