Improve navaid pronounceability (#15)
Included is a bunch of somewhat manually-tuned heuristic techniques to
transform un-pronounceable navaid names/IDs into something that a human
can pronounce and the STT lib can understand. Vosk no longer complains
that any words are not in its dictionary for KJFK navaids!

*Most* of the `:pronunciation` values were generated by these
heuristics, but a handful are just too exotic to transform
automatically, so the `atc.pronunciation.manual` may be used to manually
describe a pronunciation in such cases. When's elected, the tool will
still validate that the provided value is actually pronounceable from
STT's perspective.

As part of this change, we're making a clear distinction between `:name`
and `:pronunciation`—the former, if available, is what TTS software
should *always* use; the latter is what STT software should always
prefer for understanding humans. Previously, we did not make a
distinction, and always put the pronounceable thing in `:pronunciation`.
Making this distinction should be beneficial moving forward—especially
for navaids like `RBV` which TTS can pronounce nicely as "robbinsville"
(`:name`) but STT needs a bit more hand-holding to comprehend: "robbins
fill" (`:pronunciation`).

One of the driving forces behind these changes is that we finally
*properly* read the FST from the packed language model in order to
*precisely* get its table of known words. Doing so is actually *much*
faster than our old technique of scanning the file for a matching
string; I didn't measure it, but it's a set lookup vs a string scan with
regex so it's a pretty obvious win. Once again, our `okay` library makes
it real easy to read this structured file in, quickly.

- Improve word-pronounceability check somewhat
- Definitively fix pronunceability check by reading the FST table
- Output fixed pronunciation; distinguish from radio :name for TTS
- Fix pronounceability report
- Add some rough attempts to manipulate words into being pronounceable
- Add some more hand-tuned pronunciation tweaks
- Add some more letter swap rules
- Add one more letter swap proposal
- Try splitting small words as a last-ditch effort
- Try to treat words starting with "a" as if "a" were a separate word
- Clean up letter swaps with explanations
- Clean up pronounceability check output, fix final issues manually
- Also check :name as a way to pronounce a navaid
- Don't let STT try to use the :pronunciation key; that's for humans
- Add a small note as possible improvement