Support :runway in spawn-aircraft to start on the ground
We need a new :command for aircraft target speed to complete this.
Then, the next steps will be to 1. start the aircraft on the TWR
frequency and call in on takeoff, and 2. add the frequencies for the
adjacent airspace controllers and support handing off to them (but not
necessarily in that order)