Introduce "queues" abstraction for handling repetitive tasks
This "queues" system currently powers deparature aircraft generation,
providing endless traffic! A queue has some state, stored in the engine,
which can be checked on every tick trigger side effects on the engine,
and return a new state to be passed in on the next tick. Queues are
processed *after* aircraft simulation updates, at the very end of a tick.