Support generating infinite departing aircraft (#19)
This feature is powered by a new "queues" abstraction for handling
repetitive tasks. A queue has some state, stored in the engine, which
can
be checked on every tick trigger side effects on the engine, and return
a
new state to be passed in on the next tick. Queues are processed *after*
aircraft simulation updates, at the very end of a tick.

The departures queue uses its state to track the "next time a departure
should occur," according to the Traffic in the engine. We could consider
keeping a separate Traffic instance for arrivals and departures, using
the
"split" random feature of `test.check`, which would let us move the
Traffic
object out of the engine itself and keep it in the queue state. That in
turn could let us potentially refactor our randomness into a
stateless/immutable model, so generating new traffic would give us a new
Traffic instance, and any given instance will *always* provide the same
traffic....