Experiment with using macros to generate parsers at compile time
So far this (anecdotally) shaves off ~200ms, from 545ms to 358ms.
There's still more we can precompile; if we omit the instructions
from the on-demand compile it takes ~210ms, which still seems quite
slow. Perhaps the airport's navaids-pronounced rule can be precompiled
in the generated namespace, and then building the parsing machine is
basically just a quick map merge....