Build the grammar parser at compile time (#20)
Parsing the EBNF string grammar at runtime takes ~550ms. We originally
took that approach because there wasn't an obvious way to
programmatically generate parsing rules without using the string grammar
format, and there also wasn't an officially sanctioned way to compose
different sets of rules---which we *need* to do for airport-specific
rules (currently, just the airport's navaids).

But it turns out the grammar is essentially a map of rules, so if we're
a bit clever we can compose the rules together relatively easily. 550ms
*once* at the beginning of a session isn't that big of a deal, but it
*does* feel pretty nice!

- Experiment with using macros to generate parsers at compile time
- Precompile instructions rules
- Precompile airport context parsing data + fix rule merging, etc.
- Clean up an accomplished TODO!