Add a basic contextual "help" system, enabling navaid click for info
On left/right click of a navaid, we now show its pronunciation in the
radio history window---that should help quite a bit with learning! In
the future, we will probably prefer right-click for "help" actions like
this and left-click for "primary" actions---for example, left clicking
on an aircraft could be used to point it out to a receiving center
controller, focus its flight strip, or whatever. Right clicking, then,
can print out how to pronounce the airline, maybe expand the exit fix
name (although we might just do what openscope does and just show three
letters of the exit fix, rather than this less intuitive system).