Make *certain* sure we don't overlap arrivals at the same final fix
This doesn't necessarily help if we have multiple fixes that are
basically overlapping, but this is a pretty good algorithm. We have to
use sqrt at most 3x per aircraft so should still be quite fast, even
though we're scanning aircraft for the *furthest* from the airport.

Right now we're filtering the aircraft scanned by those arriving at the
same fix; we could eliminate overlap *for sure* by looking at *all*
arriving aircraft, but then we wouldn't have nice stacks of arrivals....