Add flight strip displays (#36)
<img width="1337" alt="Screenshot 2023-09-12 at 2 09 42 PM"
src="https://github.com/dhleong/cljs-atc/assets/816150/9b56d015-5e49-4ee8-9c5b-916424ebdf64">

There's a lot going on with this PR!

- We added a new `[browser-window]` component to abstract rendering
arbitrary content in an external window. This is pretty neat! It is
fully reactive
- We now try to disable the dark reader browser extension so colors
don't get weird
- Of course, we're rendering arrival and departure aircraft. You can pop
this out (thanks to the above component) into a separate window, for
folks with multiple monitors or something
- We're now using Google Material Symbols to render some UI icons
- We now generate unique aircraft squawk codes for each craft
- We also pick an appropriate cruise altitude for each craft
- Assigning an altitude to an aircraft will append it to a list that we
render in the strip for arriving aircraft
- All pieces of the flight strip can be right-clicked to learn what it
is