Add support for visual approaches (#39)
This PR adds some rough, initial support for clearing aircraft to fly a
visual approach!

There's certainly a lot more commands we should support to "fully"
handle visual approaches, but this includes the bare minimum "cleared
visual runway XX." Includes logic for entering the pattern, flying
downwind, turning base, and then turning final and gliding down to the
runway threshold.

Also:

- The `(tick)` and and `(command)` methods on the `ISimulated` protocol
now both accept an explicit "context." For aircraft, the context should
always be the current Engine state; for the `Engine`, we use this param
to pass in the "full" command object to `(command)` and it is unused for
`(tick)`.
- We already were passing the engine to aircraft commands in a hacky
way, using metadata on the command vector; this change makes that much
cleaner.
- This change enables command handlers (via `tick`) to cleanly query the
state of the world---for example, aircraft in a visual approach will
differ their base turn if another aircraft is currently in it.
- Adds a `:debug` traffic we can use in dev builds to quickly set up a
specific flight for testing flows. This is not available in the UI and
must be triggered from the REPL.
- Some fixes for our logic handling the final landing approach, which
should apply to the existing ILS approaches, as well.
- Organizationally, we split up the monolithic `commands.cljs` file into
a bunch of separate files in the `commands` namespace. This is because
the visual approach was quite complicated, so I wanted to give it its
own namespace. It depends on pretty much every other command to handle
navigation, etc so it was just simplest to give them their own
namespaces, as well, in order to avoid having to deal with circular
dependencies.

Does not include:

- "Correct" pattern entrance (IE: 45 degree entry to downwind at
midfield). If we ever add a "tower" gameplay mode we might want to go
back and build that, but for now the current behavior ought to be "good
enough"
- Requests to "report" positions in the pattern. Shouldn't be too hard,
but it's left to future work.
- Requests to extend the downwind. I did include logic to support such a
command in the future but again left it to future work.
- Complaining about failure to handoff to tower. In theory, the airplane
should fly missed at some point if we fail to hand it off; the same goes
for ILS approaches as well.